Apidb replication streams through pipeline.
The --replicate-apidb task used to be an all-in-one replication task
that read from the database and wrote out replication files.  It has
been enhanced to write through change streams to tasks such as the
--write-replication task.  Both ends of a replication change stream must
perform synchronisation around a ReplicationState object exchanged
during the initialization phase and persisted during the complete phase.