Added request by date support to the data server.
The replication data server has been enhanced to allow a date to be
specified in the request as well as the existing request by sequence
number.  This makes it possible for clients to request a state file
to initialise replication based on the timestamp of a data snapshot they
already have.