Fixed client restart manager to handle failures.
The sequence client restart manager was failing to recover if the retry
step failed.  It was breaking out of the retry loop upon failure.
Fixing that also resulted in a deadlock between the Netty IO thread and
monitor thread which required modification to the retry loop to only
hold the control lock while waiting for client updates after startup.