Improved build sequencing
The build scripts have been refactored to introduce several
improvements:

1. The "publish" target is never called directly.  Instead, the
"build" target includes a publish.  This is a more intuitive name
for users already familiar with ant.
2. The build-java.xml script now defines all public targets at
the end of the file for clarity.  All other targets should be
considered internal and not to be used directly.
3. The top level "all" target now completes each build stage
(ie. build, checkstyle, test) across all projects before moving
to the next stage.  Previously it would perform all steps on each
project before moving to the next project.  This improves the
likelihood of developers encountering simple checkstyle errors
before they attempt to run the full suite of tests.
4. Each phase can now be run separately across all projects.
This allows checkstyle to be run in isolation.