Generalize the merge run helper methods for all MultiSinkRunnableSources.
The methods in MergeTestUtil class were fairly generic already, but
required an EntityMerger object in the method signatures. However, the
functionality can also be reused for other similar tasks like
ChangeMerger and ChangeDeriver. This commits relaxes the type
requirements to all classes implementing MultiRunnableSinkSource and
moves the utility class to the testutil project, along with some style
and naming improvements.