Added a new PBF reader implementation.
The existing PBF reader is managed outside of Osmosis and used as
a third party library.  This library is used in several applications
and as a result it is difficult to experiment with optimisations.
A new reader task has been introduced with the aim of improving
performance.  This is currently exposed using the following task names.
--read-pbf2
--rp2
--read-pbf2-0.6