Load replication database time with timezone.
The current timestamp is loaded from the database for the purposes of
replication.  The existing query was losing the timezone portion of the
time which caused issues if the db server and client server were
running in a different timezone.  This has been fixed to use the
clock_timestamp() function which includes timezone information.