Pass custom database properties to tests
The database unit tests rely on custom system properties to define the
location of the database authentication files.  This change updates the
gradle script to pass on these custom properties from the gradle runtime
to the unit tests in a similar way to how the ant build used to do it.