Manually set some statistics for way_nodes and relation_members
Postgresql gets counts of distinct values horribly wrong for these tables.
This sometimes leads to very bad query plans, particularly for antijoin patterns

Because we have access to taginfo and know more about the data, we can do better.

Values that are likely to scale with the size of the database are set to do so.

Other values, like the maximum number of nodes in a way, are fixed based on current
data.