pgsnapshot: Remove redundant way node loading logic
The way node loading logic no longer needs to hit the way_nodes table
because the ways table has the list of node ids embedded in the nodes
column.  This logic was causing double loading until Jeffrey's recent
fix, but it is redundant and can be removed completely ... I hope.