xml: Add runtime checks for unicode support
The Xerces provided by the JDK is old and has bugs that prevent it from
parsing unicode characters that require more than 1 UTF-16 character.
The fix is to ensure a more modern version of Xerces exists on the
classpath.  This change introduces a startup check when invoking XML
parsing tasks that ensure that complex unicode characters are parsed
correctly.  If the checks fail, Osmosis fails fast.