pbf2: add way-location metadata attribute to pipeline
The primary purpose of this change is to introduce an additional
pipeline metadata attribute called "way_node.location_included" which
indicates if the PBF file contained location data on the way_node
records.

Additionally, the parsing code was re-factored considerably.  Hopefully
a little easier to comprehend now, and also does a better job of
validating that file blocks occur in the correct order.