Fix docker build to work on OSX
The Docker build was failing on OSX because the BSD tar output is
incompatible with Docker.  We use a tar file as input to docker build to
dereference symlinks which Docker itself doesn't support.  This change
avoids passing the tar file directly to Docker and instead builds a
temporary folder for the Docker context.