Remove shadow jar and jcenter
jcenter is not available from Gradle 8 so needs to be removed.
The shadow plugin is the only dependency retrieved from jcenter. I have
removed it for the moment because I don't understand why it is required
and why standard Maven dependency management can't be used to retrieve
dependencies. If it is required an alternative repo hosting it will need
to be used.