Move to Gradle Application plugin
Use the Gradle Application plugin to generate the distribution. This has
a number of advantages including much simpler Gradle configuration,
elimination of Plexus Classworlds bootstrap, more mature/tested launch
scripts.
There are some differences in the generated archives including
incorporation of a base path in all files names (probably a good thing),
elimination of /bin/extract* launch scripts (unlikely to be used any
more), and creation of a tar instead of tgz (not a major issue).